Added a claim history to the claims page to allow lectuerers to view their claims current status.
Updated verification and approvel acceptence methods to change the status of the claim on the claims table whilst also adding a new claim to their respective tables.

Updated script functions to use Ajax as i have a better understanding of it and moved the calculate function from page script to a method in the claims controller.